pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tikhonov-agent"
version = "0.1.0"
description = "Rule-based scientific agent for 1D transient inverse heat conduction via Tikhonov regularization"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26",
    "scipy>=1.12",
    "pyyaml>=6.0",
    "pandas>=2.1",
]

[project.optional-dependencies]
llm = [
    "transformers>=4.40",
    "torch>=2.2",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
